requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
argon2-cffi
python-jose
stripe
